refactor: factor out vote_count module
